+++ ChangeLog.md
@@ -17,6 +17,9 @@ Other enhancements:
 
 * `stack setup` supports installing GHC for macOS aarch64 (M1)
 
+* `stack upload` supports authentication with a Hackage API key (via
+  `HACKAGE_KEY` environment variable).
+
 Bug fixes:
 
 * Ensure that `extra-path` works for case-insensitive `PATH`s on Windows.